% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_check_that_docker_is_up.R
\name{sp_check_that_docker_is_up}
\alias{sp_check_that_docker_is_up}
\title{Check that docker is running}
\usage{
sp_check_that_docker_is_up()
}
\value{
0
}
\description{
Check that docker is running
}
\examples{
sp_check_that_docker_is_up()
}
