% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_color_markdown_text.R
\name{sp_color_markdown_text}
\alias{sp_color_markdown_text}
\title{Color subsequent text in a specified color}
\usage{
sp_color_markdown_text(x, color)
}
\arguments{
\item{x}{a string}

\item{color}{color to apply}
}
\value{
colored string
}
\description{
Color subsequent text in a specified color
}
\examples{
# sp_color_markdown_text('Cover inline tables in future section','red')
}
