% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_build}
\alias{sp_docker_build}
\title{Build a Docker image}
\usage{
sp_docker_build(build_options, build_context_path)
}
\arguments{
\item{build_options}{character: the build options. There are many; do
\code{docker build --help} at a command prompt to see them. You will need at least
the \code{--tag} option to give the built image a tag (name).}

\item{build_context_path}{character: the build context path}
}
\value{
Result of Docker command if it succeeded. Stops with an error message
if it failed.
}
\description{
Creates a Docker image
}
