% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_containers_tibble}
\alias{sp_docker_containers_tibble}
\title{List containers into a tibble}
\usage{
sp_docker_containers_tibble()
}
\value{
A tibble with all the containers
}
\description{
Creates a tibble of all containers using \code{docker ps --all}
}
\examples{
\dontrun{
sp_docker_containers_tibble()
}
}
