% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_exec}
\alias{sp_docker_exec}
\title{Execute a command in a Docker container}
\usage{
sp_docker_exec(container_name, options = "", command)
}
\arguments{
\item{container_name}{character: a valid container name. If
the container doesn't exist, the function will stop with an error
message.}

\item{options}{character: the options to use. Default is the empty
string.}

\item{command}{character: the command to run. There is no default; you
must supply a valid command.}
}
\value{
Result of Docker command if it succeeded. Stops with an error message
if it failed.
}
\description{
Executes a command in a running Docker container.
}
\details{
Do \verb{docker exec --help} in a command prompt to see all the options,
of which there are many.
}
