% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_images_tibble}
\alias{sp_docker_images_tibble}
\title{List images into a tibble}
\usage{
sp_docker_images_tibble(list_all)
}
\arguments{
\item{list_all}{logical - list all images if TRUE, just
\emph{tagged} images if FALSE}
}
\value{
A tibble listing the images
}
\description{
Creates a tibble of images using \code{docker images}
}
\examples{
\dontrun{
sp_docker_images_tibble(list_all = TRUE)
}
}
