% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_network_create}
\alias{sp_docker_network_create}
\title{Create a Docker network}
\usage{
sp_docker_network_create(network_name = "sql-pet")
}
\arguments{
\item{network_name}{character: network to create. The default is \code{sql-pet}.}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Creates a Docker network with name \code{network_name}.
}
\examples{
\dontrun{sp_docker_network_create("sql-pet")}
}
