% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_networks_tibble}
\alias{sp_docker_networks_tibble}
\title{List networks into a tibble}
\usage{
sp_docker_networks_tibble()
}
\value{
A tibble listing the networks
}
\description{
Creates a tibble of networks using \verb{docker network}
}
\examples{
\dontrun{
docker_networks <- sp_docker_networks_tibble()
View(docker_networks)
}
}
