% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_remove_container}
\alias{sp_docker_remove_container}
\title{Forcibly remove a container}
\usage{
sp_docker_remove_container(docker_container)
}
\arguments{
\item{docker_container}{character: the name of the container to
remove}
}
\value{
a numeric \code{0}
}
\description{
Forcibly removes a Docker container. If it is running it will be
forcibly terminated and removed. If it doesn't exist you won't get an error
message. This is a blunt instrument!
}
\details{
Warning: this function removes the container you asked it to remove!
}
\examples{
\dontrun{sp_docker_remove_container("sql-pet")}
}
