% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_start}
\alias{sp_docker_start}
\title{Start an existing container}
\usage{
sp_docker_start(container_name)
}
\arguments{
\item{container_name}{character container to start}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Starts the container given by \code{container_name}.
}
\examples{
\dontrun{sp_docker_start("sql-pet")}
}
