% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_stop}
\alias{sp_docker_stop}
\title{Stop an existing container}
\usage{
sp_docker_stop(container_name)
}
\arguments{
\item{container_name}{character: container to stop}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Stops the container given by \code{container_name}.
}
\examples{
\dontrun{sp_docker_stop("sql-pet")}
}
