% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_fetch_dvdrental_autodoc}
\alias{sp_fetch_dvdrental_autodoc}
\title{Fetch \code{dvdrental} autodoc}
\usage{
sp_fetch_dvdrental_autodoc(container_name = "sql-pet", file_path)
}
\arguments{
\item{container_name}{character: The container name where the \code{dvdrental}
database is running. The default is "sql-pet".}

\item{file_path}{character: A valid file path to receive the HTML file. It
should be an absolute path so you know where to find it, and the filename
should end in \code{.html} so the browser can open it. There is no default.}
}
\value{
If successful: a character value of lenght 0. If unsuccessful, the
function terminates with an error message.
}
\description{
When \code{sp_make_dvdrental_image} builds the Docker image, it
installs a utility called
\href{https://github.com/cbbrowne/autodoc}{\code{postgresql-autodoc}}. After restoring
the \code{dvdrental} database, it runs the utilily and creates an HTML file with
the database documentation on the image at
\verb{/var/lib/postgresql/dvdrental.html}. This function fetches that file to a
file on the Docker host.
}
