% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_get_dbms_data_dictionary.R
\name{sp_get_dbms_data_dictionary}
\alias{sp_get_dbms_data_dictionary}
\title{Download a table from the dbms and create a data dictionary data frame}
\usage{
sp_get_dbms_data_dictionary(table_name, con = con)
}
\arguments{
\item{table_name}{The name of a table to be processed.}

\item{con}{A DBI connection object for obtaining the local copy to be processed}
}
\value{
a data frame \code{table_dd}
}
\description{
Download a table from the dbms and create a data dictionary data frame
}
\examples{
\dontrun{all_meta <- sp_get_dbms_data_dictionary("rental")}
}
