% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_get_fivenumsum.R
\name{sp_get_fivenumsum}
\alias{sp_get_fivenumsum}
\title{Calculate a 5 number summary (for all types of variables)}
\usage{
sp_get_fivenumsum(var, character_truncate_length = 60)
}
\arguments{
\item{var}{The variable to be summarized}

\item{character_truncate_length}{the length of output character summaries (for )}
}
\value{
The min, max, median, and 25th and 75th percentiles
}
\description{
Calculate a 5 number summary (for all types of variables)
}
\examples{
summary_vector <- sp_get_fivenumsum(mtcars$mpg)
}
