% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_keyring.R
\name{sp_get_key}
\alias{sp_get_key}
\title{Get a key from the keyring}
\usage{
sp_get_key(key_name)
}
\arguments{
\item{key_name}{the name of the key to get}
}
\value{
the key_value
}
\description{
gets a key previously set into the keyring
}
\examples{
\dontrun{
sp_get_key(key_name = "POSTGRES_PASSWORD")
}
}
