% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_get_postgres_connection}
\alias{sp_get_postgres_connection}
\title{Connect to a PostgreSQL database}
\usage{
sp_get_postgres_connection(user, password, dbname, host = "localhost",
  port = "5432", seconds_to_test = 30, connection_tab = FALSE)
}
\arguments{
\item{user}{character: Username for connecting to the database. There is no
default. The user must exist and be authorized to connect.}

\item{password}{character: Password that corresponds to the username. There
is no default. The password must be correct for the user.}

\item{dbname}{character: The name of the database within the database server.
The database must exist, and the user must be authorized to access it.}

\item{host}{character: The IP address or host where postgreSQL is located,
defaults to "localhost"}

\item{port}{integer: The port on the host that postgreSQL listens to, defaults
to "5432"}

\item{seconds_to_test}{integer: The number of iterations to try while waiting
for PostgreSQL service to be ready. The function sleeps one second between
connection attempts, so a value of 10 would require approximately 10 seconds.
The default is 30.}

\item{connection_tab}{logical: \code{sp_get_postgres_connection} can open a
tab in the RStudio connections pane. \code{connection_tab = FALSE} is
the default - call with \code{connection_tab = TRUE} to open the tab.}
}
\value{
If successful: a connection object, which is an S4 object
that inherits from DBIConnection, used to communicate with the
database engine. If unsuccessful, the function terminates with an
error message.
}
\description{
Attempts to connect to a PostgreSQL database. If the connection
fails, it will retry a user-specified number of times.
}
\examples{
\dontrun{con <- sp_get_postgres_connection(
  user = "postgres",
  password = "postgres",
  dbname = "postgres"
)}
}
