% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_make_adventureworks_mssql_image}
\alias{sp_make_adventureworks_mssql_image}
\title{Make \code{adventureworks_mssql} Docker image}
\usage{
sp_make_adventureworks_mssql_image(image_tag)
}
\arguments{
\item{image_tag}{character: a valid image tag (name) for the docker image}
}
\value{
Result of Docker command if it succeeded. Stops with an error
message if it failed.
}
\description{
Creates a Docker image with Microsoft SQL Server (mssql)
and the \code{AdventureWorks2017} database
}
\examples{
\dontrun{
build_log <- sp_make_adventureworks_mssql_image(
  "adventureworks-mssql:latest"
)
sp_docker_images_tibble()
}
}
