% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_make_data_dictionary.R
\name{sp_make_data_dictionary}
\alias{sp_make_data_dictionary}
\title{Run sp_get_fivenumsum on all variables in a data frame}
\usage{
sp_make_data_dictionary(df, df_alias = NULL)
}
\arguments{
\item{df}{data frame to be summarized}

\item{df_alias}{name of the data frame to be appended}
}
\value{
a data frame
}
\description{
Run sp_get_fivenumsum on all variables in a data frame
}
\examples{
table_dd <- sp_make_data_dictionary(mtcars)
}
