% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_make_dvdrental_image.R
\name{sp_make_dvdrental_image}
\alias{sp_make_dvdrental_image}
\title{Create an image with PostgreSQL and the \code{dvdrental} database}
\usage{
sp_make_dvdrental_image(image_name)
}
\arguments{
\item{image_name}{name for the docker image}
}
\value{
Result of Docker command if it succeeded. Stops with an error message
if it failed.
}
\description{
Create an image with PostgreSQL and the \code{dvdrental} database
}
\details{
This function executes a \code{docker build} with build context of this
package's \code{extdata/docker} directory. In RStudio package development,
it's \code{inst/extdata/docker}.

The Dockerfile fetches the PostgreSQL 10 image, then installs \code{unzip}. It then
downloads and unzips the DVD rental database. Then it copyies a restore
script into the container's directory \code{/docker-entrypoint-initdb.d/restoredb.sh}.

This script will run when a container first runs the image. It simply creates
the \code{dvdrental} database and restores the unzipped \code{tar} archive to it with
\code{pg_restore}.
}
\examples{
\dontrun{sp_make_dvdrental_image("dvdrental")}
}
