% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_make_image_files.R
\name{sp_make_image_files}
\alias{sp_make_image_files}
\title{Make Image Files}
\usage{
sp_make_image_files(widget, directory = "diagrams",
  filename = "example")
}
\arguments{
\item{widget}{an `htmlwidget` object returned from DiagrammeR, usually from
`grViz` or `render_graph`}

\item{directory}{(character) a valid file path. Default is ".". If it does not
exist it will be created.}

\item{filename}{(character) a valid file name (stem only). Three files will be
overwritten if they exist and created if they don't:
\itemize{
\item <directory>/<filename>.svg
\item <directory>/<filename>.png
\item <directory>/<filename>.pdf
}}
}
\value{
png_name (character) the name of the PNG file. This is the path you
give to `knitr::include_graphics` for the `auto_pdf` option.
}
\description{
Creates image files from a DiagrammeR `htmlwidget` object
}
\details{
See the vignette "Using DiagrammeR with Bookdown" for the derivation
and some examples.
}
