% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_make_simple_pg}
\alias{sp_make_simple_pg}
\title{Make simple PostgreSQL container}
\usage{
sp_make_simple_pg(container_name)
}
\arguments{
\item{container_name}{character a valid container name for the
container}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Creates a container and runs the PostgreSQL 10
image (\code{postgres:10}) in it. The image will be downloaded if
it doesn't exist locally.
}
\examples{
\dontrun{
sp_make_simple_pg("cattle")
}
}
