% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_pg_catalog}
\alias{sp_pg_catalog}
\title{Fetch a PostgreSQL database catalog}
\usage{
sp_pg_catalog(connection)
}
\arguments{
\item{connection}{A valid open \code{DBI} connection to a PostgreSQL database.}
}
\value{
A data frame with the contents of the database catalog. The columns
are \code{schemas}, \code{name}, and \code{type}, where \code{type} is "matview" (materialized
view), "view" or "table".
}
\description{
PostgreSQL stores much of its metadata in system catalogs that
are accessible to a connected user. This function takes a connection and
returns the database catalog as a data frame.
}
\examples{
\dontrun{
library(sqlpetr)
library(dplyr)
connection <- sp_get_postgres_connection(
  user = "postgres",
  password = "postgres",
  dbname = "dvdrental"
)
print(sp_pg_catalog(connection))
}
}
