% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_pg_close_connection}
\alias{sp_pg_close_connection}
\title{Notify observer and close connection}
\usage{
sp_pg_close_connection(connection)
}
\arguments{
\item{connection}{A valid open connection from \code{sp_get_postgres_connection}.}
}
\value{
not meaningful
}
\description{
Tells the connections tab observer that connection was closed
and then closes it
}
