% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_pg_connection_closed}
\alias{sp_pg_connection_closed}
\title{Tell connections tab we closed the connection}
\usage{
sp_pg_connection_closed(connection)
}
\arguments{
\item{connection}{A valid open connection from \code{sp_get_postgres_connection}.}
}
\value{
not meaningful
}
\description{
Tells the connections tab observer that a connection was closed
}
