% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_pg_connection_opened}
\alias{sp_pg_connection_opened}
\title{Connections contract for opened connection}
\usage{
sp_pg_connection_opened(connection)
}
\arguments{
\item{connection}{A valid open connection from \code{sp_get_postgres_connection}.}
}
\value{
not meaningful
}
\description{
Registers an opened connection with the RStudio "Connections"
tab via the "Connections contract".
}
\details{
See
\url{https://rstudio.github.io/rstudio-extensions/connections-contract.html} and
\url{https://github.com/r-dbi/odbc/blob/master/R/Viewer.R}.
}
