% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_pg_databases}
\alias{sp_pg_databases}
\title{List databases in a PostgreSQL cluster}
\usage{
sp_pg_databases(connection)
}
\arguments{
\item{connection}{A valid open connection to the server}
}
\value{
A tibble listing all the databases in the cluster. This is simply
the PostgreSQL catalog \code{pg_database}. See
\url{https://www.postgresql.org/docs/10/catalog-pg-database.html}
for the documentation of the columns. \code{datname} is the database name, which
is probably all you want.
}
\description{
Connects to a PostgreSQL server and delivers a tibble with the
names of the databases in the cluster the server is hosting
}
\examples{
\dontrun{connection <- sp_get_postgres_connection(
  user = "postgres",
  password = "postgres",
  dbname = "postgres"
)
databases <- sp_pg_databases(connection)
print(databases)

}
}
