% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_pg_docker_run}
\alias{sp_pg_docker_run}
\title{Run a PostgreSQL Docker image in a container}
\usage{
sp_pg_docker_run(image_tag, container_name)
}
\arguments{
\item{image_tag}{character a valid image tag (name) for the
docker image. If it doesn't exist locally, \code{docker run} will
try to download it. If the download fails, the function will
abort.}

\item{container_name}{character a valid container name for the
container}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Creates a container and runs an image in it. The image should be based on the \code{postgres:10} image
}
\examples{
\dontrun{
sp_make_dvdrental_image("dvdrental:latest")
sp_pg_docker_run("dvdrental:latest", "sql-pet")
}
}
