% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_preview_chapter.R
\name{sp_preview_chapter}
\alias{sp_preview_chapter}
\title{Compile and browse an individual chapter}
\usage{
sp_preview_chapter(chapter_file)
}
\arguments{
\item{chapter_file}{name of the local file to preview}
}
\value{
browse
}
\description{
Compile and browse an individual chapter
}
\examples{
# sp_preview_chapter("chapter_file.Rmd")
}
