% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_print_df.R
\name{sp_print_df}
\alias{sp_print_df}
\title{Print a data frame appropriate to output type}
\usage{
sp_print_df(df)
}
\arguments{
\item{df}{the data frame to be printed}
}
\value{
either kable (in PDF) or datatable (in HTML) output
}
\description{
Print a data frame appropriate to output type
}
\examples{
# sp_print_df(df)
}
