% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_keyring.R
\name{sp_set_key}
\alias{sp_set_key}
\title{Set a key into the default keyring}
\usage{
sp_set_key(key_name, key_value)
}
\arguments{
\item{key_name}{the name of the key}

\item{key_value}{the value to set}
}
\description{
sets a key into the default keyring
}
\details{
\code{sqlpetr} uses the \code{keyring} package to manage secrets as key-value
pairs. This is portable; it will work on any platform where \code{keyring} can be
installed, including Linux, MacOS and Windows.
}
\examples{
\dontrun{
sp_set_key(key_name = "POSTGRES_PASSWORD", key_value = "postgres")
}
}
