% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for_postgres.R
\name{wait_for_postgres}
\alias{wait_for_postgres}
\title{Connect to Postgres, waiting if it is not ready}
\usage{
wait_for_postgres(user, password, dbname, seconds_to_test = 10)
}
\arguments{
\item{user}{Username for connecting to the database}

\item{password}{Password that corresponds to the username}

\item{dbname}{the name of the database within the database server}

\item{seconds_to_test}{the number of iterations to try while waiting for PostgreSQL service to be ready.}
}
\value{
When successful: a connection object, which is an S4 object that inherits from DBIConnection used to communicate with the database engine; Otherwise, an error string.
}
\description{
Assumptions: host = "localhost" and port = "5432".
}
