% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_images_tibble}
\alias{sp_docker_images_tibble}
\title{List images into a tibble}
\usage{
sp_docker_images_tibble()
}
\value{
A tibble with all the images
}
\description{
Creates a tibble of images using \code{docker images}
}
\examples{
\dontrun{
sp_docker_images_tibble()
}
}
