% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_remove_container.R
\name{sp_docker_remove_container}
\alias{sp_docker_remove_container}
\title{Removes the specified docker container}
\usage{
sp_docker_remove_container(docker_container)
}
\arguments{
\item{docker_container}{is the name of the docker container to remove.}
}
\value{
0
}
\description{
Removes the specified docker container
}
\details{
\code{sp_docker_remove_container} will forcibly remove the specified
container. If it did not exist, no errors or warnings will be given.
}
\examples{
\dontrun{sp_docker_remove_container("sql-pet")}

}
