% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_docker_run}
\alias{sp_docker_run}
\title{Run an image in a container}
\usage{
sp_docker_run(options, image, command = "", args = "")
}
\arguments{
\item{options}{the options to use when running the image. Default is the
empty string. Note that if you want to name the container, you need to at
least specify the option \code{--name}.}

\item{image}{character a valid image tag (name) for the
docker image. If it doesn't exist locally, \code{docker run} will
try to download it. If the download fails, the function will
abort.

Note that the full syntax of an image tag is \code{<repository>/<name>:<tag>}. For example, the PostgreSQL 10 image we use is \code{docker.io/postgres:10}.}

\item{command}{the command to run after the container boots up. Default is an empty string, which uses the startup command defined in the image.}

\item{args}{the arguments for the command. Default is an empty string.}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Creates a container and runs an image in it.
}
\details{
Do \code{docker run --help} in a command prompt to see all the options, of which there are many.
}
\examples{
\dontrun{
sp_docker_run(
  options = "--detach --name cattle --publish 5432:5432",
  image = "docker.io.postgres:10"
)
}
}
