% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_start.R
\name{sp_docker_start}
\alias{sp_docker_start}
\title{Start the specified docker container}
\usage{
sp_docker_start(docker_container)
}
\arguments{
\item{docker_container}{is the name of the docker container to start up.}
}
\value{
0
}
\description{
Start the specified docker container
}
\examples{
sp_docker_start("sql-pet")
}
