% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_stop.R
\name{sp_docker_stop}
\alias{sp_docker_stop}
\title{Stops the specified docker container}
\usage{
sp_docker_stop(docker_container)
}
\arguments{
\item{docker_container}{is the name of the docker container to start up.}
}
\value{
0
}
\description{
Stops the specified docker container
}
\examples{
\dontrun{sp_docker_stop("sql-pet")}
}
