% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_postgresql_functions.R
\name{sp_get_postgres_connection}
\alias{sp_get_postgres_connection}
\title{Get a DBI connection string to a PostgreSQL database, waiting if it is not ready}
\usage{
sp_get_postgres_connection(user, password, dbname, host = "localhost",
  port = "5432", seconds_to_test = 10)
}
\arguments{
\item{user}{Username for connecting to the database}

\item{password}{Password that corresponds to the username}

\item{dbname}{The name of the database within the database server}

\item{host}{The IP address or host where postgreSQL is located,
defaults to "localhost"}

\item{port}{The port on the host that postgreSQL listens to, defaults
to "5432"}

\item{seconds_to_test}{The number of iterations to try while waiting
for PostgreSQL service to be ready.}
}
\value{
If successful: a connection object, which is an S4 object
that inherits from DBIConnection, used to communicate with the
database engine. If unsuccessful, the function terminates with an
error message.
}
\description{
This function connects to an instance of PostgreSQL.
}
\examples{
\dontrun{con <- sp_get_postgres_connection("postgres", "postgres")}

}
