% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_make_dvdrental_image.R
\name{sp_make_dvdrental_image}
\alias{sp_make_dvdrental_image}
\title{Create an image with PostgreSQL and the \code{dvdrental} database}
\usage{
sp_make_dvdrental_image(image_name)
}
\arguments{
\item{image_name}{name for the docker image}
}
\value{
0
}
\description{
Create an image with PostgreSQL and the \code{dvdrental} database
}
\examples{
\dontrun{sp_make_dvdrental_image("dvdrental")}
}
