% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_docker_functions.R
\name{sp_make_dvdrental_image}
\alias{sp_make_dvdrental_image}
\title{Make \code{dvdrental} Docker image}
\usage{
sp_make_dvdrental_image(image_tag)
}
\arguments{
\item{image_tag}{character: a valid image tag (name) for the docker image}
}
\value{
Result of Docker command if it succeeded. Stops with an
error message if it failed.
}
\description{
Creates a Docker image with PostgreSQL and the \code{dvdrental}
database
}
\details{
See the vignette "Building the \code{dvdrental} Docker Image" for the details.
}
\examples{
\dontrun{
sp_make_dvdrental_image("dvdrental")
sp_docker_images_tibble()
}
}
