% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_make_simple_pg.R
\name{sp_make_simple_pg}
\alias{sp_make_simple_pg}
\title{Define and run a Docker container with PostgreSQL in it}
\usage{
sp_make_simple_pg(container_name)
}
\arguments{
\item{container_name}{name for the docker container}
}
\value{
0
}
\description{
Define and run a Docker container with PostgreSQL in it
}
\examples{
\dontrun{sp_make_simple_pg("cattle")}
}
