% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_print_df.R
\name{sp_print_df}
\alias{sp_print_df}
\title{Show (print) a data frame depending on appropriate to output type}
\usage{
sp_print_df(df)
}
\arguments{
\item{df}{the data frame to be shown}
}
\value{
either kable (in PDF) or datatable (in HTML) output or View when executing chunks in an .Rmd file
}
\description{
Show (print) a data frame depending on appropriate to output type
}
\examples{
# sp_print_df(df)
}
